backend/requirements.txt
Django==1.9.4
mysql-python
djangorestframework
django-filter